Add files via upload
The "Rock, Paper, and Scissors" project is an interactive game that allows users to play the classic hand game against the computer. It provides a fun and engaging way to test your decision-making skills and luck against an AI opponent.

Features:

User vs. Computer: The game enables the user to play against the computer, simulating the traditional "Rock, Paper, and Scissors" gameplay.

User Input: Users can make their move by choosing between three options: rock, paper, or scissors. The program will validate the user's input and respond accordingly.

Computer Random Selection: The computer opponent makes its move randomly, ensuring a fair and unpredictable gameplay experience.

Game Outcome: After both the user and computer have made their choices, the program determines the winner based on the game rules. It then displays the result of the round, indicating whether the user won, lost, or if it's a tie.

Score Tracking: The game keeps track of the user's and computer's scores throughout multiple rounds, allowing players to compete for the highest score.

Play Again: After each round, players have the option to continue playing or exit the game.

Benefits:

Provides entertainment and a nostalgic experience by recreating the popular "Rock, Paper, and Scissors" game in a digital format.
Enhances decision-making skills and strategic thinking as players try to predict and counter the opponent's moves.
Offers a simple and intuitive interface, making it accessible to players of all ages.
Provides an opportunity for friendly competition, whether playing against the computer or challenging friends to beat the high score.
Whether you're looking for a quick game to pass the time or a fun project to showcase your coding skills, the "Rock, Paper, and Scissors" project is a great choice. Enjoy the thrill of outsmarting your virtual opponent and see if you can claim the title of the ultimate "Rock, Paper, and Scissors" champion!